+++ MSF_WeaponState.cpp
@@ -34,14 +34,15 @@ ExtraWeaponState::ExtraWeaponState(ExtraRank* extraHolder)
 	this->currentState = nullptr;
 }
 
-ExtraWeaponState::WeaponState::WeaponState(UInt16 newflags, UInt16 newammoCapacity, UInt16 newchamberSize, UInt16 newshotCount, UInt64 newloadedAmmo, TESAmmo* newchamberedAmmo, std::vector<TESAmmo*>* newBCRammo, std::vector<BGSMod::Attachment::Mod*>* newStateMods)
-{
+ExtraWeaponState::WeaponState::WeaponState(UInt16 newflags, UInt16 newammoCapacity, UInt16 newchamberSize, UInt32 newchamberedCount, UInt16 newshotCount, UInt64 newloadedAmmo, TESAmmo* newchamberedAmmo, TESAmmo* newequippedAmmo, std::vector<TESAmmo*>* newBCRammo, std::vector<BGSMod::Attachment::Mod*>* newStateMods) {
 	this->flags = newflags;
 	this->ammoCapacity = newammoCapacity;
 	this->chamberSize = newchamberSize;
+	this->chamberedCount = newchamberedCount;
 	this->shotCount = newshotCount;
 	this->loadedAmmo = newloadedAmmo;
 	this->chamberedAmmo = newchamberedAmmo;
+	this->equippedAmmo = newequippedAmmo;
 	if (newBCRammo)
 		this->BCRammo = *newBCRammo;
 	if (newStateMods)
@@ -101,9 +102,11 @@ ExtraWeaponState::WeaponState* ExtraWeaponState::WeaponState::Clone()
 	newState->flags = this->flags;
 	newState->ammoCapacity = this->ammoCapacity;
 	newState->chamberSize = this->chamberSize;
+	newState->chamberedCount = this->chamberedCount;
 	newState->shotCount = this->shotCount;
 	newState->loadedAmmo = this->loadedAmmo;
 	newState->chamberedAmmo = this->chamberedAmmo;
+	newState->equippedAmmo = this->equippedAmmo;
 	newState->BCRammo = this->BCRammo;
 	newState->stateMods = this->stateMods;
 	return newState;
@@ -125,6 +128,8 @@ bool ExtraWeaponState::WeaponState::FillData(ExtraDataList* extraDataList, Equip
 	this->ammoCapacity = 0;
 	this->flags = 0;
 	this->shotCount = 0;
+	this->chamberedCount = 0;
+	this->equippedAmmo = nullptr;
 	this->BCRammo.clear();
 	this->stateMods.clear();
 	//has BCR?
@@ -147,14 +152,18 @@ bool ExtraWeaponState::WeaponState::FillData(ExtraDataList* extraDataList, Equip
 		//BGSMod::Attachment::Mod* receiver = Utilities::GetModAtAttachPoint(attachedMods, MSF_MainData::receiverAP);
 		MSF_Data::GetChamberData(attachedMods, currInstanceData, &this->chamberSize, &this->flags);
 		this->chamberedAmmo = currInstanceData->ammo;
+		this->equippedAmmo = currInstanceData->ammo;
 		this->ammoCapacity = currInstanceData->ammoCapacity;
 		if (!equipData)
 			this->loadedAmmo = this->ammoCapacity;
-
+		this->chamberedCount = this->loadedAmmo < this->chamberSize ? this->loadedAmmo : this->chamberSize;
+		if (MSF_Data::InstanceHasTRSupport(currInstanceData))
+			this->flags |= bHasTacticalReload;
 		if (MSF_Data::InstanceHasBCRSupport(currInstanceData))
 		{
+			this->flags |= bHasBCR;
 			for (UInt32 ammoIdx = 0; ammoIdx < this->loadedAmmo; ammoIdx++)
-				this->BCRammo.push_back(this->chamberedAmmo);
+				this->BCRammo.push_back(this->equippedAmmo);
 		}
 	}
 
@@ -169,7 +178,7 @@ bool ExtraWeaponState::WeaponState::FillData(ExtraDataList* extraDataList, Equip
 	return true;
 }
 
-bool ExtraWeaponState::HandleWeaponStateEvents(UInt8 eventType, Actor* actor)
+bool ExtraWeaponState::HandleWeaponStateEvents(UInt8 eventType, Actor* actor, UInt8 eventSubtype, UInt32 oldLoadedAmmoCount)
 {
 	if (!(MSF_MainData::MCMSettingFlags & MSF_MainData::mMakeExtraRankMask))
 		return true;
@@ -199,7 +208,7 @@ bool ExtraWeaponState::HandleWeaponStateEvents(UInt8 eventType, Actor* actor)
 	break;
 	case ExtraWeaponState::kEventTypeAmmoCount: ret = weaponState->HandleAmmoChangeEvent(equippedWeapExtraData, equippedData); break;
 	case ExtraWeaponState::kEventTypeFireWeapon: ret = weaponState->HandleFireEvent(equippedWeapExtraData, equippedData); break;
-	case ExtraWeaponState::kEventTypeReload: ret = weaponState->HandleReloadEvent(equippedWeapExtraData, equippedData, eventType); break;
+	case ExtraWeaponState::kEventTypeReload: ret = weaponState->HandleReloadEvent(equippedWeapExtraData, equippedData, eventSubtype, oldLoadedAmmoCount); break;
 	//case ExtraWeaponState::KEventTypeModded: ret = weaponState->HandleModChangeEvent(equippedWeapExtraData, equippedData); break;
 	default:
 		return false; break;
@@ -331,10 +340,13 @@ bool ExtraWeaponState::WeaponState::UpdateAmmoState(ExtraDataList* extraDataList
 			modsToModify->push_back(std::pair<BGSMod::Attachment::Mod*, bool>(targetAmmoMod, toAttach));
 	}
 	if (eventType == kEventTypeModdedSwitch)
+	{
 		this->chamberedAmmo = finalAmmo; //this->currentState->chamberedAmmo = instanceData->ammo;
+		this->equippedAmmo = finalAmmo;
+	}
 	else if (eventType == kEventTypeModdedGameplay && this->chamberedAmmo != finalAmmo)
 	{
-		this->loadedAmmo = this->ammoCapacity;
+		this->loadedAmmo = this->ammoCapacity; //!!!
 		if (hasBCR)
 		{
 			this->BCRammo.clear();
@@ -342,6 +354,7 @@ bool ExtraWeaponState::WeaponState::UpdateAmmoState(ExtraDataList* extraDataList
 				this->BCRammo.push_back(finalAmmo);
 		}
 		this->chamberedAmmo = finalAmmo;
+		this->equippedAmmo = finalAmmo;
 	}
 		
 	return true;
@@ -435,25 +448,25 @@ bool ExtraWeaponState::HandleEquipEvent(ExtraDataList* extraDataList, EquipWeapo
 
 bool ExtraWeaponState::HandleFireEvent(ExtraDataList* extraDataList, EquipWeaponData* equipData)
 {
-	//if (!this->currentState)
-	//	this->UpdateWeaponStates(extraDataList, equipData);
-	//_DEBUG("fireWeaponState");
-	//if (this->currentState->switchToAmmoAfterFire)
-	//{
-	//	this->currentState->shotCount++;
-	//	if (this->currentState->shotCount >= this->currentState->chamberSize)
-	//	{
-	//		//evaluate
-	//		if (MSF_Base::AttachModToEquippedWeapon(*g_player, this->currentState->switchToAmmoAfterFire->mod, true, 0, false))
-	//		{
-	//			this->currentState->currentSwitchedAmmo = this->currentState->switchToAmmoAfterFire;
-	//			this->currentState->switchToAmmoAfterFire = nullptr;
-	//			this->currentState->shotCount = 0;
-	//			MSF_Scaleform::UpdateWidgetData();
-	//		}
-	//		//error handle
-	//	}
-	//}
+	_DEBUG("fireWeaponState");
+	if (!this->currentState)
+		this->UpdateWeaponStates(extraDataList, equipData, kEventTypeReload);
+	else
+	{
+		if (this->currentState->BCRammo.size() > 0)
+		{
+			UInt32 idxDel = 0;
+			UInt32 idxNext = 0;
+			if (this->currentState->flags & ExtraWeaponState::WeaponState::bChamberLIFO)
+			{
+				idxDel = this->currentState->BCRammo.size() - 1;
+				idxNext = this->currentState->BCRammo.size() - 2;
+			}
+			TESAmmo* firedAmmo = this->currentState->BCRammo[idxDel];
+			TESAmmo* nextAmmo = this->currentState->BCRammo[idxNext];
+			this->currentState->BCRammo.erase(this->currentState->BCRammo.begin() + idxDel);
+		}
+	}
 	return true;
 }
 
@@ -462,7 +475,9 @@ bool ExtraWeaponState::HandleAmmoChangeEvent(ExtraDataList* extraDataList, Equip
 	if (!this->currentState)
 		this->UpdateWeaponStates(extraDataList, equipData, kEventTypeAmmoCount);
 	this->currentState->loadedAmmo = equipData->loadedAmmoCount;
-	//_DEBUG("ammoChangeState");
+	if (this->currentState->loadedAmmo < this->currentState->chamberSize)
+		this->currentState->chamberedCount = this->currentState->loadedAmmo;
+	_DEBUG("ammoChangeState");
 	//if (equipData->loadedAmmoCount == 0 && this->currentState->switchToAmmoAfterFire)
 	//{
 	//	//evaluate
@@ -478,22 +493,34 @@ bool ExtraWeaponState::HandleAmmoChangeEvent(ExtraDataList* extraDataList, Equip
 	return true;
 }
 
-bool ExtraWeaponState::HandleReloadEvent(ExtraDataList* extraDataList, EquipWeaponData* equipData, UInt8 eventType) //0:start,1:BCR,2:complete
+bool ExtraWeaponState::HandleReloadEvent(ExtraDataList* extraDataList, EquipWeaponData* equipData, UInt8 eventType, UInt32 oldLoadedAmmoCount) //0:start,1:BCR,2:complete
 {
-	//_DEBUG("reloadState");
-	//ExtraInstanceData* extraInstanceData = DYNAMIC_CAST(extraDataList->GetByType(kExtraData_InstanceData), BSExtraData, ExtraInstanceData);
-	//TESObjectWEAP* baseWeap = DYNAMIC_CAST(extraInstanceData->baseForm, TESForm, TESObjectWEAP);
-	//BGSObjectInstanceExtra* attachedMods = DYNAMIC_CAST(extraDataList->GetByType(kExtraData_ObjectInstance), BSExtraData, BGSObjectInstanceExtra);
-	//TESObjectWEAP::InstanceData* currInstanceData = (TESObjectWEAP::InstanceData*)Runtime_DynamicCast(extraInstanceData->instanceData, RTTI_TBO_InstanceData, RTTI_TESObjectWEAP__InstanceData);
-	//if (!currInstanceData || !baseWeap)
-	//	return false;
-	//if (!this->currentState)
-	//	this->UpdateWeaponStates(extraDataList, equipData);
-	//if (this->currentState->flags & ExtraWeaponState::WeaponState::bHasTacticalReload)
-	//{
-	//	equipData->loadedAmmoCount = currInstanceData->ammoCapacity + this->currentState->chamberSize; //if chamberedRounds < chamberSize ?
-	//	this->currentState->loadedAmmo = equipData->loadedAmmoCount;
-	//}
+	_DEBUG("reloadState");
+	ExtraInstanceData* extraInstanceData = DYNAMIC_CAST(extraDataList->GetByType(kExtraData_InstanceData), BSExtraData, ExtraInstanceData);
+	TESObjectWEAP* baseWeap = DYNAMIC_CAST(extraInstanceData->baseForm, TESForm, TESObjectWEAP);
+	TESObjectWEAP::InstanceData* currInstanceData = (TESObjectWEAP::InstanceData*)Runtime_DynamicCast(extraInstanceData->instanceData, RTTI_TBO_InstanceData, RTTI_TESObjectWEAP__InstanceData);
+	if (!currInstanceData || !baseWeap)
+		return false;
+	if (this->currentState)
+	{
+		UInt16 BCRtype = MSF_MainData::modSwitchManager.GetIsBCRreload();
+		if (BCRtype)
+		{
+			if (MSF_Data::InstanceHasTRSupport(currInstanceData))
+				MSF_MainData::BCRinterfaceHolder.SetBCRammoCap(this->currentState->ammoCapacity + 1);
+			this->currentState->BCRammo.push_back(this->currentState->equippedAmmo);
+			this->currentState->chamberedCount = equipData->loadedAmmoCount < this->currentState->chamberSize ? equipData->loadedAmmoCount : this->currentState->chamberSize;
+		}
+		else
+		{
+			UInt32 ammoToUse = Utilities::GetInventoryItemCount((*g_player)->inventoryList, this->currentState->equippedAmmo);
+			if (ammoToUse > currInstanceData->ammoCapacity)
+				ammoToUse = currInstanceData->ammoCapacity;
+			if (MSF_Data::InstanceHasTRSupport(currInstanceData))
+				equipData->loadedAmmoCount = ammoToUse + (oldLoadedAmmoCount < this->currentState->chamberedCount ? oldLoadedAmmoCount : this->currentState->chamberedCount); //currInstanceData->ammoCapacity +
+			this->currentState->chamberedCount = equipData->loadedAmmoCount < this->currentState->chamberSize ? equipData->loadedAmmoCount : this->currentState->chamberSize;
+		}
+	}
 
 	return true;
 }
@@ -622,6 +649,34 @@ TESAmmo* ExtraWeaponState::GetCurrentAmmo()
 		return nullptr;
 }
 
+bool ExtraWeaponState::SetEquippedAmmo(TESAmmo* ammo)
+{
+	if (this->currentState && ammo)
+		return this->currentState->equippedAmmo = ammo;
+	else
+		return false;
+}
+
+TESAmmo* ExtraWeaponState::GetEquippedAmmo()
+{
+	if (this->currentState)
+		return this->currentState->equippedAmmo;
+	else
+		return nullptr;
+}
+
+ExtraWeaponState::AmmoStateData* ExtraWeaponState::GetAmmoStateData()
+{
+	if (!this->currentState)
+		return nullptr;
+	AmmoStateData data;
+	data.ammoCapacity = this->currentState->ammoCapacity;
+	data.chamberedCount = this->currentState->chamberedCount;
+	data.chamberSize = this->currentState->chamberSize;
+	data.loadedAmmo = this->currentState->loadedAmmo;
+	return &data;
+}
+
 void ExtraWeaponState::PrintStoredData()
 {
 	_MESSAGE("ID: %08X; total states: %08X; active state marked with *", this->ID, this->weaponStates.size());
@@ -634,12 +689,15 @@ void ExtraWeaponState::PrintStoredData()
 		if (state.second)
 		{
 			UInt32 ammoFormID = 0;
+			UInt32 chAmmoFormID = 0;
 			if (state.second->chamberedAmmo)
-				ammoFormID = state.second->chamberedAmmo->formID;
+				chAmmoFormID = state.second->chamberedAmmo->formID;
+			if (state.second->equippedAmmo)
+				ammoFormID = state.second->equippedAmmo->formID;
 			if (state.second == this->currentState)
-				_MESSAGE("-State*%02X: mod: %08X; chamberedAmmo: %08X; loadedAmmoCount %i; chamberSize: %i; ammoCapacity: %i; shotCount: %i; flags: %04X", idx, modFormID, ammoFormID, state.second->loadedAmmo, state.second->chamberSize, state.second->ammoCapacity, state.second->shotCount, state.second->flags);
+				_MESSAGE("-State*%02X: mod: %08X; chamberedAmmo: %08X; equippedAmmo: %08X; loadedAmmoCount %i; chamberSize: %i; chamberedCount: %i; ammoCapacity: %i; shotCount: %i; flags: %04X", idx, modFormID, chAmmoFormID, ammoFormID, state.second->loadedAmmo, state.second->chamberSize, state.second->chamberedCount, state.second->ammoCapacity, state.second->shotCount, state.second->flags);
 			else
-				_MESSAGE("-State %02X: mod: %08X; chamberedAmmo: %08X; loadedAmmoCount %i; chamberSize: %i; ammoCapacity: %i; shotCount: %i; flags: %04X", idx, modFormID, ammoFormID, state.second->loadedAmmo, state.second->chamberSize, state.second->ammoCapacity, state.second->shotCount, state.second->flags);
+				_MESSAGE("-State %02X: mod: %08X; chamberedAmmo: %08X; equippedAmmo: %08X; loadedAmmoCount %i; chamberSize: %i; chamberedCount: %i; ammoCapacity: %i; shotCount: %i; flags: %04X", idx, modFormID, chAmmoFormID, ammoFormID, state.second->loadedAmmo, state.second->chamberSize, state.second->chamberedCount, state.second->ammoCapacity, state.second->shotCount, state.second->flags);
 		}
 		else if (state.second == this->currentState)
 			_MESSAGE("-State*%02X: mod: %08X", idx, modFormID);